refactor(aipanel): 拆出工作区标签外壳，内层面板迁至 ai/AIConversationTabPanel
- 新增 components/ai/useAIWorkspaceTabs.ts：标签组持久化与订阅、待定位消息队列、关闭防误触锁、滚动同步、标签新建/激活/关闭/分叉与跨终端打开
- 新增 components/ai/AIWorkspaceTabBar.tsx：标签栏渲染段（溢出滚动/执行中圆点/右键菜单/新建按钮）
- 新增 components/ai/AIConversationTabPanel.tsx：单个标签页对话面板编排层（组合各领域 hook 并分发渲染段）
- 新增 components/ai/AIConversationPanelSections.tsx：输入区（AIComposer）与设置浮层渲染段
- AIPanel.tsx 1160 -> 64 行（仅保留外壳）
