+++ frontend/src/components/ai/useAIWorkspaceTabs.ts
@@ -1,0 +1,557 @@
+import { useCallback, useEffect, useRef, useState } from 'react'
+import type * as React from 'react'
+import { cancelAIChat } from './aiChatBridge.ts'
+import { getAIConversation, saveAIConversation, subscribeAIConversationChanges } from './aiConversationBridge.ts'
+import { AI_WORKSPACE_TAB_CLOSE_QUIET_MS, type AIPanelProps } from './aiChatLogic.ts'
+import {
+  clearAIWorkspaceTabPendingLocation,
+  createAIWorkspaceTabId,
+  findAIWorkspaceConversationTab,
+  getAIWorkspaceTabGroup,
+  getAIWorkspaceTabPendingLocation,
+  setAIWorkspaceTabGroup,
+  setAIWorkspaceTabPendingLocation,
+  subscribeAIWorkspaceTabGroup,
+  type AIWorkspaceTab,
+  type AIWorkspaceTabGroup,
+} from '../../utils/aiWorkspaceTabs.ts'
+import { type I18nKey } from '../../i18n.ts'
+
+type LooseT = (key: I18nKey, vars?: Record<string, unknown>) => string
+
+// AI 工作区多标签管理 hook：标签组的持久化与订阅、待定位消息队列、关闭防误触锁、
+// 滚动状态同步、标签新建/激活/关闭/分叉、跨终端打开会话与标签运行时状态回报。
+// 从 AIPanel 外壳原样搬移，闭包依赖经参数同名注入，代码零改动。
+export function useAIWorkspaceTabs({ t, terminalId, sessionId, onActiveTabChange, onActivateWorkspaceTab }: {
+  t: LooseT
+  terminalId: string
+  sessionId: string
+  onActiveTabChange?: (tabId: string) => void
+  onActivateWorkspaceTab?: (terminalId: string, tabId: string) => void
+}) {
+  const [tabGroup, setTabGroup] = useState<AIWorkspaceTabGroup>(() => getAIWorkspaceTabGroup(terminalId))
+  const [tabRequestIds, setTabRequestIds] = useState<Record<string, string>>({})
+  const [aiWorkspaceTabOverflow, setAIWorkspaceTabOverflow] = useState(false)
+  const [aiWorkspaceTabCanScrollLeft, setAIWorkspaceTabCanScrollLeft] = useState(false)
+  const [aiWorkspaceTabCanScrollRight, setAIWorkspaceTabCanScrollRight] = useState(false)
+  const tabGroupRef = useRef(tabGroup)
+  const tabRequestIdsRef = useRef<Record<string, string>>({})
+  const aiWorkspaceTabRuntimeRef = useRef<Record<string, { conversationId: string; activeRequestId: string }>>({})
+  const aiWorkspaceTabScrollRef = useRef<HTMLDivElement | null>(null)
+  const aiWorkspaceTabCloseLockRef = useRef<{ tabId: string; confirmed: boolean; lastInteractionAt: number } | undefined>(undefined)
+  const aiWorkspaceTabCloseUnlockTimerRef = useRef<number | undefined>(undefined)
+  useEffect(() => {
+    tabGroupRef.current = tabGroup
+  }, [tabGroup])
+  useEffect(() => subscribeAIWorkspaceTabGroup(terminalId, setTabGroup), [terminalId])
+  useEffect(() => {
+    const tabIds = new Set(tabGroup.tabs.map((tab) => tab.id))
+    const nextRuntime: Record<string, { conversationId: string; activeRequestId: string }> = {}
+    tabIds.forEach((tabId) => {
+      const runtime = aiWorkspaceTabRuntimeRef.current[tabId]
+      if (runtime) {
+        nextRuntime[tabId] = runtime
+      }
+    })
+    aiWorkspaceTabRuntimeRef.current = nextRuntime
+    setTabRequestIds((current) => {
+      const next = Object.fromEntries(Object.entries(current).filter(([tabId]) => tabIds.has(tabId)))
+      return Object.keys(next).length === Object.keys(current).length ? current : next
+    })
+  }, [tabGroup.tabs])
+  useEffect(() => {
+    onActiveTabChange?.(tabGroup.activeTabId)
+  }, [onActiveTabChange, tabGroup.activeTabId])
+  const updateTabGroup = useCallback((updater: (current: AIWorkspaceTabGroup) => AIWorkspaceTabGroup) => {
+    return setAIWorkspaceTabGroup(terminalId, updater)
+  }, [terminalId])
+  const flushAIWorkspaceTabPendingLocation = useCallback((tabId: string) => {
+    const pendingLocation = getAIWorkspaceTabPendingLocation(terminalId, tabId)
+    const runtime = aiWorkspaceTabRuntimeRef.current[tabId]
+    const tab = getAIWorkspaceTabGroup(terminalId).tabs.find((item) => item.id === tabId)
+    if (
+      !pendingLocation
+      || !runtime
+      || runtime.conversationId !== pendingLocation.conversationId
+      || tab?.conversationId !== pendingLocation.conversationId
+      || typeof window === 'undefined'
+    ) {
+      return
+    }
+    window.requestAnimationFrame(() => {
+      window.requestAnimationFrame(() => {
+        const currentPendingLocation = getAIWorkspaceTabPendingLocation(terminalId, tabId)
+        const currentRuntime = aiWorkspaceTabRuntimeRef.current[tabId]
+        const currentTab = getAIWorkspaceTabGroup(terminalId).tabs.find((item) => item.id === tabId)
+        if (
+          currentPendingLocation?.conversationId !== pendingLocation.conversationId
+          || currentPendingLocation?.messageId !== pendingLocation.messageId
+          || !currentRuntime
+          || currentRuntime.conversationId !== pendingLocation.conversationId
+          || currentTab?.conversationId !== pendingLocation.conversationId
+        ) {
+          return
+        }
+        clearAIWorkspaceTabPendingLocation(terminalId, tabId)
+        window.dispatchEvent(new CustomEvent('ai-conversation-diff-locate', {
+          detail: {
+            sessionId,
+            terminalId,
+            tabId,
+            messageId: pendingLocation.messageId,
+          },
+        }))
+      })
+    })
+  }, [sessionId, terminalId])
+  const queueAIWorkspaceTabLocation = useCallback((targetTerminalId: string, tabId: string, conversationId: string, messageId: string) => {
+    const normalizedMessageId = typeof messageId === 'string' ? messageId.trim() : ''
+    if (!targetTerminalId || !tabId || !conversationId || !normalizedMessageId) {
+      return
+    }
+    setAIWorkspaceTabPendingLocation(targetTerminalId, tabId, {
+      conversationId,
+      messageId: normalizedMessageId,
+    })
+    if (targetTerminalId === terminalId) {
+      flushAIWorkspaceTabPendingLocation(tabId)
+    }
+  }, [flushAIWorkspaceTabPendingLocation, terminalId])
+  useEffect(() => {
+    tabGroup.tabs.forEach((tab) => flushAIWorkspaceTabPendingLocation(tab.id))
+  }, [flushAIWorkspaceTabPendingLocation, tabGroup.tabs])
+  const clearAIWorkspaceTabCloseUnlockTimer = useCallback(() => {
+    if (aiWorkspaceTabCloseUnlockTimerRef.current !== undefined) {
+      window.clearTimeout(aiWorkspaceTabCloseUnlockTimerRef.current)
+      aiWorkspaceTabCloseUnlockTimerRef.current = undefined
+    }
+  }, [])
+  const scheduleAIWorkspaceTabCloseUnlock = useCallback(() => {
+    clearAIWorkspaceTabCloseUnlockTimer()
+    const schedule = () => {
+      const closeLock = aiWorkspaceTabCloseLockRef.current
+      if (!closeLock?.confirmed) {
+        return
+      }
+      const remainingDelay = Math.max(0, AI_WORKSPACE_TAB_CLOSE_QUIET_MS - (Date.now() - closeLock.lastInteractionAt))
+      aiWorkspaceTabCloseUnlockTimerRef.current = window.setTimeout(() => {
+        aiWorkspaceTabCloseUnlockTimerRef.current = undefined
+        const currentCloseLock = aiWorkspaceTabCloseLockRef.current
+        if (!currentCloseLock?.confirmed) {
+          return
+        }
+        if (Date.now() - currentCloseLock.lastInteractionAt < AI_WORKSPACE_TAB_CLOSE_QUIET_MS) {
+          schedule()
+          return
+        }
+        aiWorkspaceTabCloseLockRef.current = undefined
+      }, remainingDelay)
+    }
+    schedule()
+  }, [clearAIWorkspaceTabCloseUnlockTimer])
+  const suppressAIWorkspaceTabCloseInteraction = useCallback((event: React.SyntheticEvent) => {
+    const closeLock = aiWorkspaceTabCloseLockRef.current
+    if (!closeLock) {
+      return
+    }
+    closeLock.lastInteractionAt = Date.now()
+    event.preventDefault()
+    event.stopPropagation()
+    scheduleAIWorkspaceTabCloseUnlock()
+  }, [scheduleAIWorkspaceTabCloseUnlock])
+  const syncAIWorkspaceTabScrollState = useCallback(() => {
+    const element = aiWorkspaceTabScrollRef.current
+    if (!element) {
+      setAIWorkspaceTabOverflow(false)
+      setAIWorkspaceTabCanScrollLeft(false)
+      setAIWorkspaceTabCanScrollRight(false)
+      return
+    }
+    const maxScrollLeft = Math.max(0, element.scrollWidth - element.clientWidth)
+    const hasOverflow = maxScrollLeft > 1
+    setAIWorkspaceTabOverflow(hasOverflow)
+    setAIWorkspaceTabCanScrollLeft(hasOverflow && element.scrollLeft > 1)
+    setAIWorkspaceTabCanScrollRight(hasOverflow && element.scrollLeft < maxScrollLeft - 1)
+  }, [])
+  const scrollActiveAIWorkspaceTabIntoView = useCallback((tabId: string) => {
+    const element = aiWorkspaceTabScrollRef.current
+    if (!element || !tabId) {
+      return
+    }
+    const tabElement = Array.from(element.querySelectorAll<HTMLElement>('[data-ai-workspace-tab-id]'))
+      .find((item) => item.dataset.aiWorkspaceTabId === tabId)
+    if (!tabElement) {
+      return
+    }
+    const scrollRect = element.getBoundingClientRect()
+    const tabRect = tabElement.getBoundingClientRect()
+    const edgePadding = 6
+    const delta = tabRect.left < scrollRect.left + edgePadding
+      ? tabRect.left - scrollRect.left - edgePadding
+      : tabRect.right > scrollRect.right - edgePadding
+        ? tabRect.right - scrollRect.right + edgePadding
+        : 0
+    if (delta) {
+      element.scrollBy({ left: delta, behavior: 'smooth' })
+    }
+  }, [])
+  const scrollAIWorkspaceTabs = useCallback((direction: number) => {
+    const element = aiWorkspaceTabScrollRef.current
+    if (!element) {
+      return
+    }
+    const step = Math.max(96, Math.round(element.clientWidth * 0.45))
+    element.scrollBy({ left: step * direction, behavior: 'smooth' })
+  }, [])
+  const handleAIWorkspaceTabScroll = useCallback(() => {
+    syncAIWorkspaceTabScrollState()
+  }, [syncAIWorkspaceTabScrollState])
+  const handleAIWorkspaceTabWheel = useCallback((event: React.WheelEvent<HTMLDivElement>) => {
+    const element = aiWorkspaceTabScrollRef.current
+    if (!element || element.scrollWidth <= element.clientWidth) {
+      return
+    }
+    const delta = Math.abs(event.deltaX) > Math.abs(event.deltaY) ? event.deltaX : event.deltaY
+    if (!delta) {
+      return
+    }
+    element.scrollBy({ left: delta, behavior: 'auto' })
+    event.preventDefault()
+  }, [])
+  useEffect(() => {
+    if (tabGroup.tabs.length > 0) {
+      if (!tabGroup.tabs.some((tab) => tab.id === tabGroup.activeTabId)) {
+        updateTabGroup((current) => ({
+          ...current,
+          activeTabId: current.tabs[0]?.id || '',
+        }))
+      }
+      return
+    }
+    const tabId = createAIWorkspaceTabId()
+    updateTabGroup((current) => (
+      current.tabs.length > 0
+        ? current
+        : {
+            activeTabId: tabId,
+            tabs: [{ id: tabId, conversationId: '', title: t('新对话'), transient: false }],
+          }
+    ))
+  }, [t, tabGroup.activeTabId, tabGroup.tabs, updateTabGroup])
+  useEffect(() => {
+    const closeLock = aiWorkspaceTabCloseLockRef.current
+    if (!closeLock || tabGroup.tabs.some((tab) => tab.id === closeLock.tabId)) {
+      return
+    }
+    closeLock.confirmed = true
+    scheduleAIWorkspaceTabCloseUnlock()
+  }, [scheduleAIWorkspaceTabCloseUnlock, tabGroup.tabs])
+  useEffect(() => () => clearAIWorkspaceTabCloseUnlockTimer(), [clearAIWorkspaceTabCloseUnlockTimer])
+  useEffect(() => {
+    const frameId = window.requestAnimationFrame(() => {
+      scrollActiveAIWorkspaceTabIntoView(tabGroup.activeTabId)
+      syncAIWorkspaceTabScrollState()
+    })
+    return () => window.cancelAnimationFrame(frameId)
+  }, [scrollActiveAIWorkspaceTabIntoView, syncAIWorkspaceTabScrollState, tabGroup.activeTabId, tabGroup.tabs])
+  useEffect(() => {
+    const element = aiWorkspaceTabScrollRef.current
+    if (!element) {
+      return undefined
+    }
+    const handleResize = () => {
+      scrollActiveAIWorkspaceTabIntoView(tabGroup.activeTabId)
+      syncAIWorkspaceTabScrollState()
+    }
+    const observer = typeof ResizeObserver === 'function' ? new ResizeObserver(handleResize) : null
+    observer?.observe(element)
+    window.addEventListener('resize', handleResize)
+    handleResize()
+    return () => {
+      observer?.disconnect()
+      window.removeEventListener('resize', handleResize)
+    }
+  }, [scrollActiveAIWorkspaceTabIntoView, syncAIWorkspaceTabScrollState, tabGroup.activeTabId])
+  const dismissTabPanels = useCallback((tabId: string) => {
+    if (typeof window === 'undefined' || !tabId) {
+      return
+    }
+    window.dispatchEvent(new CustomEvent('ai-change-review-clear', {
+      detail: { sessionId: terminalId, tabId },
+    }))
+    window.dispatchEvent(new CustomEvent('ai-conversation-diff-close', {
+      detail: { sessionId, terminalId, tabId },
+    }))
+  }, [sessionId, terminalId])
+  const createWorkspaceTab = useCallback(() => {
+    const tabId = createAIWorkspaceTabId()
+    updateTabGroup((current) => ({
+      activeTabId: tabId,
+      tabs: [...current.tabs, { id: tabId, conversationId: '', title: t('新对话'), transient: false }],
+    }))
+    return tabId
+  }, [t, updateTabGroup])
+  const returnWorkspaceTabHome = useCallback((tabId: string) => {
+    updateTabGroup((current) => ({
+      ...current,
+      tabs: current.tabs.map((tab) => (
+        tab.id === tabId
+          ? { ...tab, conversationId: '', title: t('新对话'), transient: false }
+          : tab
+      )),
+    }))
+  }, [t, updateTabGroup])
+  const activateWorkspaceTab = useCallback((tabId: string) => {
+    updateTabGroup((current) => (
+      current.tabs.some((tab) => tab.id === tabId)
+        ? { ...current, activeTabId: tabId }
+        : current
+    ))
+  }, [updateTabGroup])
+  const closeWorkspaceTab = useCallback((tabId: string) => {
+    const current = tabGroupRef.current
+    const tabIndex = current.tabs.findIndex((tab) => tab.id === tabId)
+    if (tabIndex < 0 || current.tabs.length <= 1) {
+      return
+    }
+    const requestId = tabRequestIdsRef.current[tabId]
+    delete tabRequestIdsRef.current[tabId]
+    delete aiWorkspaceTabRuntimeRef.current[tabId]
+    clearAIWorkspaceTabPendingLocation(terminalId, tabId)
+    setTabRequestIds((currentRequests) => {
+      if (!currentRequests[tabId]) {
+        return currentRequests
+      }
+      const nextRequests = { ...currentRequests }
+      delete nextRequests[tabId]
+      return nextRequests
+    })
+    if (requestId) {
+      void cancelAIChat(requestId)
+    }
+    dismissTabPanels(tabId)
+    updateTabGroup((group) => {
+      const index = group.tabs.findIndex((tab) => tab.id === tabId)
+      if (index < 0) {
+        return group
+      }
+      const tabs = group.tabs.filter((tab) => tab.id !== tabId)
+      const activeTabId = group.activeTabId === tabId
+        ? (tabs[Math.max(0, index - 1)]?.id || '')
+        : group.activeTabId
+      return { activeTabId, tabs }
+    })
+  }, [dismissTabPanels, updateTabGroup])
+  const forkWorkspaceTabConversation = useCallback(async (sourceConversationId: string, sourceTabId: string, openInNewTab: boolean) => {
+    const normalizedSourceId = typeof sourceConversationId === 'string' ? sourceConversationId.trim() : ''
+    if (!normalizedSourceId) {
+      return
+    }
+    let forkedId = ''
+    try {
+      const snapshot = await getAIConversation(normalizedSourceId)
+      if (!snapshot?.id) {
+        return
+      }
+      const baseTitle = typeof snapshot.title === 'string' && snapshot.title.trim() ? snapshot.title.trim() : t('新对话')
+      const now = Date.now()
+      const forkedSnapshot = await saveAIConversation({
+        ...snapshot,
+        id: '',
+        title: `${baseTitle} - ${t('副本')}`,
+        parentConversationId: '',
+        rootConversationId: '',
+        relationType: '',
+        relationSource: '',
+        parentTitleSnapshot: '',
+        archived: false,
+        status: 'idle',
+        createdAt: now,
+        updatedAt: now,
+      })
+      forkedId = typeof forkedSnapshot?.id === 'string' ? forkedSnapshot.id.trim() : ''
+    } catch {
+      return
+    }
+    if (!forkedId) {
+      return
+    }
+    if (openInNewTab) {
+      const newTabId = createAIWorkspaceTabId()
+      updateTabGroup((current) => ({
+        activeTabId: newTabId,
+        tabs: [...current.tabs, { id: newTabId, conversationId: forkedId, title: '' }],
+      }))
+      return
+    }
+    const normalizedSourceTabId = typeof sourceTabId === 'string' ? sourceTabId.trim() : ''
+    updateTabGroup((current) => {
+      if (!normalizedSourceTabId || !current.tabs.some((tab) => tab.id === normalizedSourceTabId)) {
+        const newTabId = createAIWorkspaceTabId()
+        return {
+          activeTabId: newTabId,
+          tabs: [...current.tabs, { id: newTabId, conversationId: forkedId, title: '' }],
+        }
+      }
+      return {
+        activeTabId: normalizedSourceTabId,
+        tabs: current.tabs.map((tab) => (
+          tab.id === normalizedSourceTabId
+            ? { ...tab, conversationId: forkedId, title: '' }
+            : tab
+        )),
+      }
+    })
+  }, [t, updateTabGroup])
+  const openConversationInWorkspaceTab = useCallback(async (conversationId: string, messageId = '') => {
+    const normalizedConversationId = typeof conversationId === 'string' ? conversationId.trim() : ''
+    if (!normalizedConversationId) {
+      return
+    }
+    const existing = findAIWorkspaceConversationTab(normalizedConversationId)
+    if (existing) {
+      queueAIWorkspaceTabLocation(existing.terminalId, existing.tabId, normalizedConversationId, messageId)
+      if (existing.terminalId !== terminalId) {
+        setAIWorkspaceTabGroup(existing.terminalId, (current) => ({
+          ...current,
+          activeTabId: existing.tabId,
+        }))
+        onActivateWorkspaceTab?.(existing.terminalId, existing.tabId)
+        return
+      }
+      activateWorkspaceTab(existing.tabId)
+      return
+    }
+    const activeTabId = tabGroupRef.current.activeTabId
+    const activeTabExists = tabGroupRef.current.tabs.some((tab) => tab.id === activeTabId)
+    const activeRequestId = activeTabExists
+      ? (tabRequestIdsRef.current[activeTabId] || tabRequestIds[activeTabId] || '')
+      : ''
+    const tabId = activeTabExists && !activeRequestId ? activeTabId : createAIWorkspaceTabId()
+    updateTabGroup((current) => ({
+      activeTabId: tabId,
+      tabs: current.tabs.some((tab) => tab.id === tabId)
+        ? current.tabs.map((tab) => (
+            tab.id === tabId
+              ? { ...tab, conversationId: normalizedConversationId, title: '', transient: false }
+              : tab
+          ))
+        : [...current.tabs, {
+            id: tabId,
+            conversationId: normalizedConversationId,
+            title: '',
+            transient: false,
+          }],
+    }))
+    queueAIWorkspaceTabLocation(terminalId, tabId, normalizedConversationId, messageId)
+  }, [activateWorkspaceTab, onActivateWorkspaceTab, queueAIWorkspaceTabLocation, tabRequestIds, terminalId, updateTabGroup])
+  const handleWorkspaceTabStateChange = useCallback((tabId: string, state: { conversationId: string; title: string; activeRequestId: string; transient: boolean }) => {
+    const conversationId = state.transient ? '' : state.conversationId
+    aiWorkspaceTabRuntimeRef.current[tabId] = {
+      conversationId,
+      activeRequestId: state.activeRequestId,
+    }
+    if (state.activeRequestId) {
+      tabRequestIdsRef.current[tabId] = state.activeRequestId
+    } else {
+      delete tabRequestIdsRef.current[tabId]
+    }
+    setTabRequestIds((currentRequests) => {
+      const currentRequestId = currentRequests[tabId] || ''
+      if (currentRequestId === state.activeRequestId) {
+        return currentRequests
+      }
+      if (!state.activeRequestId) {
+        const nextRequests = { ...currentRequests }
+        delete nextRequests[tabId]
+        return nextRequests
+      }
+      return {
+        ...currentRequests,
+        [tabId]: state.activeRequestId,
+      }
+    })
+    const title = state.transient ? (state.title || t('临时会话')) : (conversationId ? state.title : t('新对话'))
+    const currentTab = tabGroupRef.current.tabs.find((tab) => tab.id === tabId)
+    flushAIWorkspaceTabPendingLocation(tabId)
+    if (!currentTab || (currentTab.conversationId === conversationId && currentTab.title === title && currentTab.transient === state.transient)) {
+      return
+    }
+    updateTabGroup((current) => ({
+      ...current,
+      tabs: current.tabs.map((tab) => (
+        tab.id === tabId
+          ? { ...tab, conversationId, title, transient: state.transient }
+          : tab
+      )),
+    }))
+  }, [flushAIWorkspaceTabPendingLocation, t, updateTabGroup])
+  useEffect(() => subscribeAIConversationChanges((change: unknown) => {
+    const detail = change && typeof change === 'object' ? change as Record<string, unknown> : null
+    const conversationId = typeof detail?.conversationId === 'string' ? detail.conversationId.trim() : ''
+    if (detail?.type !== 'delete' || !conversationId) {
+      return
+    }
+    const affectedTabs = tabGroupRef.current.tabs.filter((tab) => tab.conversationId === conversationId)
+    affectedTabs.forEach((tab) => {
+      const requestId = tabRequestIdsRef.current[tab.id]
+      delete tabRequestIdsRef.current[tab.id]
+      delete aiWorkspaceTabRuntimeRef.current[tab.id]
+      clearAIWorkspaceTabPendingLocation(terminalId, tab.id)
+      if (requestId) {
+        void cancelAIChat(requestId)
+      }
+      dismissTabPanels(tab.id)
+    })
+    if (affectedTabs.length === 0) {
+      return
+    }
+    const affectedTabIds = new Set(affectedTabs.map((tab) => tab.id))
+    setTabRequestIds((currentRequests) => Object.fromEntries(
+      Object.entries(currentRequests).filter(([tabId]) => !affectedTabIds.has(tabId)),
+    ))
+    updateTabGroup((current) => {
+      const tabs = current.tabs.filter((tab) => !affectedTabIds.has(tab.id))
+      return {
+        activeTabId: affectedTabIds.has(current.activeTabId) ? (tabs[0]?.id || '') : current.activeTabId,
+        tabs,
+      }
+    })
+  }), [dismissTabPanels, updateTabGroup])
+  const activeTabId = tabGroup.activeTabId
+  return {
+    tabGroup,
+    setTabGroup,
+    tabRequestIds,
+    setTabRequestIds,
+    aiWorkspaceTabOverflow,
+    aiWorkspaceTabCanScrollLeft,
+    aiWorkspaceTabCanScrollRight,
+    tabGroupRef,
+    tabRequestIdsRef,
+    aiWorkspaceTabRuntimeRef,
+    aiWorkspaceTabScrollRef,
+    aiWorkspaceTabCloseLockRef,
+    aiWorkspaceTabCloseUnlockTimerRef,
+    updateTabGroup,
+    flushAIWorkspaceTabPendingLocation,
+    queueAIWorkspaceTabLocation,
+    clearAIWorkspaceTabCloseUnlockTimer,
+    scheduleAIWorkspaceTabCloseUnlock,
+    suppressAIWorkspaceTabCloseInteraction,
+    syncAIWorkspaceTabScrollState,
+    scrollActiveAIWorkspaceTabIntoView,
+    scrollAIWorkspaceTabs,
+    handleAIWorkspaceTabScroll,
+    handleAIWorkspaceTabWheel,
+    dismissTabPanels,
+    createWorkspaceTab,
+    returnWorkspaceTabHome,
+    activateWorkspaceTab,
+    closeWorkspaceTab,
+    forkWorkspaceTabConversation,
+    openConversationInWorkspaceTab,
+    handleWorkspaceTabStateChange,
+    activeTabId,
+  }
+}
